packages/manugen-ai/pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "manugen-ai"
version = "0.1.0"
description = ""
readme = "README.md"
license = "BSD-3-Clause"
authors = [
  { name = "Dave Bunten", email = "david.bunten@cuanschutz.edu" },
  { name = "Vincent Rubinetti", email = "vincent.rubinetti@cuanschutz.edu" },
  { name = "Milton Pividori", email = "milton.pividori@cuanschutz.edu" },
  { name = "Faisal Alquaddoomi", email = "faisal.alquaddoomi@cuanschutz.edu" },
]
requires-python = ">=3.12,<3.13"
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "cyclopts>=3.17,<4",
  "duckdb>=1.3.1",
  "flagembedding>=1.3.5",
  "google-adk>=1.2.1,<2",
  "ipython>=9.2",
  "jsonschema>=4.24",
  "litellm>=1.72.1",
  "nbconvert>=7.16.6",
  "pillow>=11.2.1",
  "poethepoet>=0.35",
  "psycopg2-binary>=2.9.10",
  "pyalex>=0.18",
  "pygit2>=1.18",
  "python-dotenv>=1.1",
  "requests>=2.32.3",
  "transformers>=4.52.4",
]
scripts.manugen = "manugen_ai.cli:app"

[dependency-groups]
dev = [
  "black>=25.1,<26",
  "ipywidgets>=8.1.7",
  "jupyterlab>=4.4.3",
  "nbclient>=0.10.2",
  "nbformat>=5.10.4",
  "pre-commit>=4.2.0",
  "pytest>=8.4,<9",
  "pytest-asyncio>=1",
]

[tool.hatch.build.targets.sdist]
include = [ "src/manugen_ai" ]

[tool.hatch.build.targets.wheel]
include = [ "src/manugen_ai" ]

[tool.hatch.build.targets.wheel.sources]
"src/manugen_ai" = "manugen_ai"

[tool.poe.tasks]
# note: quarto commands below expect quarto installed on the local system.
# see here for more information: https://quarto.org/docs/download/
create_withdrarxiv_embeddings.shell = """
python -c \
"from manugen_ai.data import create_withdrarxiv_embeddings; \
create_withdrarxiv_embeddings()"
"""
